+++ Changes.md
@@ -24,7 +24,15 @@ Breaking Changes
 1.5.x.x (relative to 1.5.10.1)
 =======
 
+Improvements
+------------
 
+- RenderMan : Stylized Looks no longer require manual AOV setup. The relevant AOVs are added automatically whenever a stylized display filter is present.
+
+Fixes
+-----
+
+- RenderManShader : Fixed handling of minimum and maximum values for `color`, `vector`, `normal` and `point` parameters.
 
 1.5.10.1 (relative to 1.5.10.0)
 ========